+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - allow to customize AIDER_DESK_DATA_DIR via environment variable
 - added support for Azure provider
 - fixed not being possible to interrupt agent run when switching between projects
+- fixed issue with context files not being highlighted in All files mode on Windows
 
 ## [0.32.0]